+++ src/realm/descriptor.hpp
@@ -62,6 +62,12 @@ namespace _impl { class DescriptorFriend; }
 /// Note, however, that this equivalence holds only as long as no
 /// shared subtable descriptors are involved.
 ///
+/// Root descriptor is created by table which also keeps a shared ref
+/// on it. Sub descriptor is created by root descriptor which also keeps
+/// a shared ref on it. This means for most cases, the descriptors live
+/// as long as the root table. Don't hold a shared ref on it to avoid
+/// circular refs.
+///
 /// \sa Table::get_descriptor()
 class Descriptor {
 public:
@@ -314,10 +320,13 @@ public:
     /// If this descriptor is the *root descriptor*, then this
     /// function returns null. Otherwise it returns the accessor of
     /// the parent descriptor.
+    /// The raw pointer of the parent descriptor lives as long as this descriptor
+    /// and it is not owned by this descriptor. So don't try to free the returned
+    /// pointer and there is no need to free it.
     ///
     /// \sa is_root()
-    DescriptorRef get_parent() noexcept;
-    ConstDescriptorRef get_parent() const noexcept;
+    Descriptor* get_parent() noexcept;
+    const Descriptor* get_parent() const noexcept;
     //@}
 
     //@{
@@ -408,15 +417,35 @@ public:
     /// debugging purposes.
     size_t get_num_unique_values(size_t column_ndx) const;
 
+
+    // Public for make_shared. Use DescriptorFriend::create() instead for creating new
+    // instances.
+    //
+    // Called by the root table to create the root descriptor.
+    //
+    // Puts this descriptor accessor into the attached state. This
+    // attaches it to the underlying structure of array nodes. When
+    // this function returns, is_attached() will return true.
+    //
+    // Not idempotent.
+    //
+    // The specified table is not allowed to be a subtable with a
+    // shareable spec. That is, Table::has_shared_spec() must return
+    // false.
+    //
+    // The specified spec must be the spec of the specified table.
+    //
+    // The specified spec here is not owned by the created descriptor,
+    // which means it won't be freed here.
+    Descriptor(Table* table, Spec* spec) noexcept;
+
     ~Descriptor() noexcept;
 
 
 private:
-    TableRef m_root_table; // Table associated with root descriptor. Detached iff null.
-    DescriptorRef m_parent; // Null iff detached or root descriptor.
-    Spec* m_spec; // Valid if attached. Owned iff valid and `m_parent`.
-
-    mutable unsigned long m_ref_count;
+    TableRef m_root_table; // Table associated with root descriptor. Detached if null.
+    Descriptor* const m_parent; // Null if detached or root descriptor. Not owned by this.
+    Spec* m_spec; // Valid if attached. Owned if valid and `m_parent`.
 
     // Whenever a subtable descriptor accessor is created, it is
     // stored in this map. This ensures that when get_subdescriptor()
@@ -428,20 +457,18 @@ private:
     // objects.
     struct subdesc_entry {
         size_t m_column_ndx;
-        Descriptor* m_subdesc;
-        subdesc_entry(size_t column_ndx, Descriptor*);
+        DescriptorRef m_subdesc;
+        subdesc_entry(size_t column_ndx, DescriptorRef& desc_ref) noexcept;
     };
     typedef std::vector<subdesc_entry> subdesc_map;
+    // No lock is needed for this map to guarantee the destructor's thread safety.
+    // If it is a root destructor, the map will be cleared and there won't be
+    // concurrent operations on the map anymore.
+    // If it is a sub destructor, no operation on its parent's map.
     mutable subdesc_map m_subdesc_map;
 
-    Descriptor() noexcept;
-
-    void bind_ptr() const noexcept;
-    void unbind_ptr() const noexcept;
-
-    // Called by the root table if this becomes the root
-    // descriptor. Otherwise it is called by the descriptor that
-    // becomes its parent.
+    // Called by the descriptor that becomes its parent to create a
+    // sub descriptor.
     //
     // Puts this descriptor accessor into the attached state. This
     // attaches it to the underlying structure of array nodes. It does
@@ -455,17 +482,14 @@ private:
     // shareable spec. That is, Table::has_shared_spec() must return
     // false.
     //
-    // The specified spec must be the spec of the specified table or
-    // of one of its direct or indirect subtable columns.
+    // The specified spec must be the spec of one of its direct or
+    // indirect subtable columns.
     //
-    // When the specified spec is the spec of the root table, the
-    // parent must be specified as null. When the specified spec is
-    // not the root spec, a proper parent must be specified.
-    void attach(Table*, Descriptor* parent, Spec*) noexcept;
+    // The created descriptor owns the specified spec which means it
+    // will free it.
+    Descriptor(Table* table, Descriptor* parent, Spec* spec) noexcept;
 
-    // Detach accessor from underlying descriptor. Caller must ensure
-    // that a reference count exists upon return, for example by
-    // obtaining an extra reference count before the call.
+    // Detach accessor from underlying descriptor.
     //
     // This function is called either by the root table if this is the
     // root descriptor, or by the parent descriptor, if it is not.
@@ -475,6 +499,10 @@ private:
     // also calls detach_subdesc_accessors(). When this function
     // returns, is_attached() will return false.
     //
+    // This won't remove its entry from subdesc_map if it is called
+    // on a sub descriptor. Thus this is not supposed to be called
+    // outside to detach a sub descriptor.
+    //
     // Not idempotent.
     void detach() noexcept;
 
@@ -482,10 +510,6 @@ private:
     // exist, that is, all subtable descriptor accessors that have
     // this descriptor as ancestor.
     void detach_subdesc_accessors() noexcept;
-
-    // Remove the entry from m_subdesc_map that refers to the
-    // specified subtable descriptor. It must be there.
-    void remove_subdesc_entry(Descriptor* subdesc) const noexcept;
 
     // Record the path in terms of subtable column indexes from the
     // root descriptor to this descriptor. If this descriptor is a
@@ -498,7 +522,10 @@ private:
     // Returns a pointer to the accessor of the specified
     // subdescriptor if that accessor exists, otherwise this function
     // return null.
-    Descriptor* get_subdesc_accessor(size_t column_ndx) noexcept;
+    // The parent descriptor creates and owns the sub descriptor. Don't hold
+    // a shared ref to the returned descriptor if the caller has been holding
+    // a TableRef already.
+    DescriptorRef get_subdesc_accessor(size_t column_ndx) noexcept;
 
     void move_column(size_t from_ndx, size_t to_ndx);
 
@@ -506,8 +533,6 @@ private:
     void adj_erase_column(size_t col_ndx) noexcept;
     void adj_move_column(size_t col_ndx_1, size_t col_ndx_2) noexcept;
 
-    friend class util::bind_ptr<Descriptor>;
-    friend class util::bind_ptr<const Descriptor>;
     friend class _impl::DescriptorFriend;
 };
 
@@ -667,14 +692,14 @@ inline ConstDescriptorRef Descriptor::get_subdescriptor(size_t column_ndx) const
     return const_cast<Descriptor*>(this)->get_subdescriptor(column_ndx);
 }
 
-inline DescriptorRef Descriptor::get_parent() noexcept
+inline Descriptor* Descriptor::get_parent() noexcept
 {
     return m_parent;
 }
 
-inline ConstDescriptorRef Descriptor::get_parent() const noexcept
+inline const Descriptor* Descriptor::get_parent() const noexcept
 {
-    return const_cast<Descriptor*>(this)->get_parent();
+    return m_parent;
 }
 
 inline TableRef Descriptor::get_root_table() noexcept
@@ -706,38 +731,30 @@ inline bool Descriptor::is_root() const noexcept
     return !m_parent;
 }
 
-inline Descriptor::Descriptor() noexcept: m_ref_count(0)
+inline Descriptor::Descriptor(Table* table, Spec* spec) noexcept
+:m_parent(nullptr), m_spec(spec)
 {
-}
-
-inline void Descriptor::bind_ptr() const noexcept
-{
-    ++m_ref_count;
-}
-
-inline void Descriptor::unbind_ptr() const noexcept
-{
-    if (--m_ref_count == 0)
-        delete this;
-}
-
-inline void Descriptor::attach(Table* table, Descriptor* parent, Spec* spec) noexcept
-{
-    REALM_ASSERT(!is_attached());
     REALM_ASSERT(!table->has_shared_type());
     m_root_table.reset(table);
-    m_parent.reset(parent);
-    m_spec = spec;
 }
+
+inline Descriptor::Descriptor(Table* table, Descriptor* parent, Spec* spec) noexcept
+:m_parent(parent), m_spec(spec)
+{
+    REALM_ASSERT(parent);
+    REALM_ASSERT(!table->has_shared_type());
+    m_root_table.reset(table);
+}
+
 
 inline bool Descriptor::is_attached() const noexcept
 {
     return bool(m_root_table);
 }
 
-inline Descriptor::subdesc_entry::subdesc_entry(size_t n, Descriptor* d):
+inline Descriptor::subdesc_entry::subdesc_entry(size_t n, DescriptorRef& desc_ref) noexcept:
     m_column_ndx(n),
-    m_subdesc(d)
+    m_subdesc(desc_ref)
 {
 }
 
@@ -757,21 +774,15 @@ inline bool Descriptor::operator!=(const Descriptor& d) const noexcept
 // not all of the non-public parts of the Descriptor class.
 class _impl::DescriptorFriend {
 public:
-    static Descriptor* create()
+    static Descriptor* create(Table* table, Spec* spec) noexcept
     {
-        return new Descriptor; // Throws
-    }
-
-    static void attach(Descriptor& desc, Table* table, Descriptor* parent, Spec* spec) noexcept
-    {
-        desc.attach(table, parent, spec);
+        return new Descriptor(table, spec);
     }
 
     static void detach(Descriptor& desc) noexcept
     {
         desc.detach();
     }
-
     static Table& get_root_table(Descriptor& desc) noexcept
     {
         return *desc.m_root_table;
@@ -798,7 +809,7 @@ public:
         return desc.record_subdesc_path(begin, end);
     }
 
-    static Descriptor* get_subdesc_accessor(Descriptor& desc, size_t column_ndx) noexcept
+    static DescriptorRef get_subdesc_accessor(Descriptor& desc, size_t column_ndx) noexcept
     {
         return desc.get_subdesc_accessor(column_ndx);
     }