+++ src/realm/descriptor_fwd.hpp
@@ -19,14 +19,13 @@
 #ifndef REALM_DESCRIPTOR_FWD_HPP
 #define REALM_DESCRIPTOR_FWD_HPP
 
-#include <realm/util/bind_ptr.hpp>
-
+#include <memory>
 
 namespace realm {
 
 class Descriptor;
-typedef util::bind_ptr<Descriptor> DescriptorRef;
-typedef util::bind_ptr<const Descriptor> ConstDescriptorRef;
+using DescriptorRef = std::shared_ptr<Descriptor>;
+using ConstDescriptorRef = std::shared_ptr<const Descriptor>;
 
 } // namespace realm